+++ projects/packages/my-jetpack/_inc/hooks/use-analytics/index.js
@@ -1,4 +1,3 @@
-/* global myJetpackInitialState */
 /**
  * External dependencies
  */
@@ -7,9 +6,8 @@ import jetpackAnalytics from '@automattic/jetpack-analytics';
 import useMyJetpackConnection from '../use-my-jetpack-connection';
 
 const useAnalytics = () => {
-	const { isUserConnected, userConnectionData = {} } = useMyJetpackConnection();
+	const { isUserConnected, connectedPlugins, userConnectionData = {} } = useMyJetpackConnection();
 	const { login, ID } = userConnectionData.currentUser?.wpcomUser || {};
-	const { connectedPlugins } = myJetpackInitialState;
 
 	/**
 	 * Initialize tracks with user data.
@@ -45,16 +43,14 @@ const useAnalytics = () => {
 	 * @param {string} event       - event name
 	 * @param {object} properties  - event propeties
 	 */
-	const recordMyJetpackEvent = useCallback(
-		( event, properties ) => {
-			tracks.recordEvent( event, {
-				...properties,
-				version: window?.myJetpackInitialState?.myJetpackVersion,
-				referring_plugins: connectedPluginsSlugs,
-			} );
-		},
-		[ tracks, connectedPluginsSlugs ]
-	);
+	const recordMyJetpackEvent = useCallback( ( event, properties ) => {
+		tracks.recordEvent( event, {
+			...properties,
+			version: window?.myJetpackInitialState?.myJetpackVersion,
+			referring_plugins: connectedPluginsSlugs,
+		} );
+		// eslint-disable-next-line react-hooks/exhaustive-deps
+	}, [] );
 
 	return {
 		clearedIdentity,