+++ projects/packages/my-jetpack/_inc/state/actions.js
@@ -3,6 +3,7 @@
  */
 import apiFetch from '@wordpress/api-fetch';
 import { __, sprintf } from '@wordpress/i18n';
+import { CONNECTION_STORE_ID } from '@automattic/jetpack-connection';
 
 /**
  * Internal dependencies
@@ -81,9 +82,10 @@ function setIsFetchingProduct( productId, isFetching ) {
  * @param {object}   store          - Redux store.
  * @param {object}   store.select   - Redux store select.
  * @param {Function} store.dispatch - Redux store dispatch.
+ * @param {object}   store.registry - Redux registry.
  * @returns {Promise}               - Promise which resolves when the product status is updated.
  */
-function requestProductStatus( productId, data, { select, dispatch } ) {
+function requestProductStatus( productId, data, { select, dispatch, registry } ) {
 	return new Promise( ( resolve, reject ) => {
 		// Check valid product.
 		const isValid = select.isValidProduct( productId );
@@ -109,6 +111,7 @@ function requestProductStatus( productId, data, { select, dispatch } ) {
 			.then( freshProduct => {
 				dispatch( setIsFetchingProduct( productId, false ) );
 				dispatch( setProduct( freshProduct ) );
+				registry.dispatch( CONNECTION_STORE_ID ).refreshConnectedPlugins();
 				resolve( freshProduct?.status );
 			} )
 			.catch( error => {