+++ projects/packages/waf/tests/php/test-waf-standalone-bootstrap.php
@@ -5,7 +5,7 @@
  * @package automattic/jetpack-waf
  */
 
-use Automattic\Jetpack\Waf\WafStandaloneBootstrap;
+use Automattic\Jetpack\Waf\Waf_Standalone_Bootstrap;
 
 /**
  * Runtime test suite.
@@ -14,11 +14,13 @@ final class WafStandaloneBootstrapTest extends PHPUnit\Framework\TestCase {
 
 	/**
 	 * Test guarding against running outside of WP context.
+	 *
+	 * @runInSeparateProcess
 	 */
 	public function testConstructingTheBootstrapWithoutAbspathConstantThrowsException() {
 		$this->assertFalse( defined( 'ABSPATH' ) );
 		$this->expectExceptionMessage( 'Cannot generate the WAF bootstrap if we are not running in WordPress context.' );
-		new WafStandaloneBootstrap();
+		new Waf_Standalone_Bootstrap();
 	}
 
 	/**
@@ -31,10 +33,12 @@ final class WafStandaloneBootstrapTest extends PHPUnit\Framework\TestCase {
 		define( 'WP_CONTENT_DIR', '/pseudo/dir' );
 
 		$this->assertFalse( defined( 'JETPACK_WAF_DIR' ) );
+		$this->assertFalse( defined( 'JETPACK_WAF_WPCONFIG' ) );
 
-		new WafStandaloneBootstrap();
+		new Waf_Standalone_Bootstrap();
 
 		$this->assertSame( '/pseudo/dir/jetpack-waf', JETPACK_WAF_DIR );
+		$this->assertSame( '/pseudo/dir/../wp-config.php', JETPACK_WAF_WPCONFIG );
 	}
 
 	/**
@@ -46,7 +50,7 @@ final class WafStandaloneBootstrapTest extends PHPUnit\Framework\TestCase {
 		define( 'ABSPATH', '/pseudo' );
 		define( 'WP_CONTENT_DIR', '/pseudo/dir' );
 
-		$mock_builder = $this->getMockBuilder( WafStandaloneBootstrap::class );
+		$mock_builder = $this->getMockBuilder( Waf_Standalone_Bootstrap::class );
 		$mock_builder->setMethods( array( 'initialize_filesystem' ) );
 
 		$sut = $mock_builder->getMock();
@@ -93,7 +97,7 @@ final class WafStandaloneBootstrapTest extends PHPUnit\Framework\TestCase {
 		global $wp_filesystem;
 		$wp_filesystem = $filesystem_mock;
 
-		$mock_builder = $this->getMockBuilder( WafStandaloneBootstrap::class );
+		$mock_builder = $this->getMockBuilder( Waf_Standalone_Bootstrap::class );
 		$mock_builder->setMethods( array( 'initialize_filesystem' ) );
 
 		$sut = $mock_builder->getMock();
@@ -131,7 +135,7 @@ final class WafStandaloneBootstrapTest extends PHPUnit\Framework\TestCase {
 		global $wp_filesystem;
 		$wp_filesystem = $filesystem_mock;
 
-		$mock_builder = $this->getMockBuilder( WafStandaloneBootstrap::class );
+		$mock_builder = $this->getMockBuilder( Waf_Standalone_Bootstrap::class );
 		$mock_builder->setMethods( array( 'initialize_filesystem' ) );
 
 		$sut = $mock_builder->getMock();
@@ -170,7 +174,7 @@ final class WafStandaloneBootstrapTest extends PHPUnit\Framework\TestCase {
 		global $wp_filesystem;
 		$wp_filesystem = $filesystem_mock;
 
-		$mock_builder = $this->getMockBuilder( WafStandaloneBootstrap::class );
+		$mock_builder = $this->getMockBuilder( Waf_Standalone_Bootstrap::class );
 		$mock_builder->setMethods( array( 'initialize_filesystem' ) );
 
 		$sut = $mock_builder->getMock();
@@ -205,7 +209,7 @@ final class WafStandaloneBootstrapTest extends PHPUnit\Framework\TestCase {
 		global $wp_filesystem;
 		$wp_filesystem = $filesystem_mock;
 
-		$mock_builder = $this->getMockBuilder( WafStandaloneBootstrap::class );
+		$mock_builder = $this->getMockBuilder( Waf_Standalone_Bootstrap::class );
 		$mock_builder->setMethods( array( 'initialize_filesystem' ) );
 
 		$sut = $mock_builder->getMock();