+++ projects/plugins/boost/app/features/optimizations/cloud-css/Cloud_CSS_Request.php
@@ -10,28 +10,15 @@
 namespace Automattic\Jetpack_Boost\Features\Optimizations\Cloud_CSS;
 
 use Automattic\Jetpack_Boost\Lib\Boost_API;
-use Automattic\Jetpack_Boost\Lib\Critical_CSS\Critical_CSS_State;
 use Automattic\Jetpack_Boost\Lib\Critical_CSS\Critical_CSS_Storage;
 
 /**
- * Cloud CSS State
+ * Cloud CSS Request
  */
 class Cloud_CSS_Request {
-
-	/**
-	 * @var Critical_CSS_State
-	 */
-	private $state;
-
-	public function __construct( $state ) {
-		$this->state = $state;
-	}
-
-	public function request_generate() {
-		$sources = $this->state->get_provider_urls();
-
+	public function request_generate( $providers ) {
 		$client               = Boost_API::get_client();
-		$payload              = array( 'providers' => $sources );
+		$payload              = array( 'providers' => $providers );
 		$payload['requestId'] = md5( wp_json_encode( $payload ) );
 
 		$this->reset_existing_css();