+++ projects/plugins/jetpack/extensions/shared/components/product-management-controls/utils.js
@@ -69,6 +69,30 @@ const messages = {
 			'jetpack'
 		),
 	},
+	'invalid product configured for this block': {
+		[ PRODUCT_TYPE_PAYMENT_PLAN ]: __(
+			'Invalid payment plan configured for this block.',
+			'jetpack'
+		),
+		[ PRODUCT_TYPE_SUBSCRIPTION ]: __(
+			'Invalid subscription configured for this block.',
+			'jetpack'
+		),
+	},
+	'the button will be hidden from your visitors until you select a valid product': {
+		[ PRODUCT_TYPE_PAYMENT_PLAN ]: __(
+			'The button will be hidden from your visitors until you select a valid payment plan.',
+			'jetpack'
+		),
+		[ PRODUCT_TYPE_SUBSCRIPTION ]: __(
+			'The subscribe button will be hidden from your visitors until you select a valid subscription.',
+			'jetpack'
+		),
+	},
+	'mark this product as a donation': {
+		[ PRODUCT_TYPE_PAYMENT_PLAN ]: __( 'Mark this payment plan as a donation', 'jetpack' ),
+		[ PRODUCT_TYPE_SUBSCRIPTION ]: __( 'Mark this subscription as a donation', 'jetpack' ),
+	},
 };
 
 export function getMessageByProductType( message, productType = PRODUCT_TYPE_PAYMENT_PLAN ) {