+++ projects/plugins/jetpack/tests/php/json-api/test-class.json-api-plugins-endpoints.php
@@ -80,7 +80,6 @@ class WP_Test_Jetpack_Json_Api_Plugins_Endpoints extends WP_UnitTestCase {
 			mkdir( $the_real_folder );
 			$clean = true;
 		}
-// phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_read_file_put_contents
 		file_put_contents(
 			$the_real_file,
 			'<?php
@@ -154,7 +153,6 @@ class WP_Test_Jetpack_Json_Api_Plugins_Endpoints extends WP_UnitTestCase {
 			$clean = true;
 		}
 
-		// phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_read_file_put_contents
 		file_put_contents(
 			$the_real_file,
 			'<?php
@@ -238,7 +236,6 @@ class WP_Test_Jetpack_Json_Api_Plugins_Endpoints extends WP_UnitTestCase {
 			$clean = true;
 		}
 
-		// phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_read_file_put_contents
 		file_put_contents(
 			$the_real_file,
 			'<?php
@@ -333,11 +330,11 @@ class WP_Test_Jetpack_Json_Api_Plugins_Endpoints extends WP_UnitTestCase {
 		$this->rmdir( $the_real_folder );
 	}
 
-	function filesystem_method_direct( $method ) {
+	public function filesystem_method_direct() {
 		return 'direct';
 	}
 
-	function rmdir( $dir ) {
+	public function rmdir( $dir ) {
 		foreach ( scandir( $dir ) as $file ) {
 			if ( is_dir( $file ) ) {
 				continue;