+++ projects/plugins/jetpack/tests/php/modules/publicize/test_class.publicize.php
@@ -126,7 +126,7 @@ class WP_Test_Publicize extends WP_UnitTestCase {
 	}
 
 	public function test_filter_can_prevent_publicize() {
-		add_filter( 'publicize_should_publicize_published_post', array( $this, 'prevent_publicize_post' ), 10, 2 );
+		add_filter( 'publicize_should_publicize_published_post', array( $this, 'prevent_publicize_post' ) );
 
 		$this->post->post_status = 'publish';
 
@@ -151,7 +151,7 @@ class WP_Test_Publicize extends WP_UnitTestCase {
 		unregister_post_type( 'foo' );
 	}
 
-	function assertPublicized( $should_have_publicized, $post ) {
+	public function assertPublicized( $should_have_publicized, $post ) {
 		if ( $should_have_publicized ) {
 			$this->assertEquals( $post->ID, $this->publicized_post_id, 'Is not the same post ID' );
 			$this->assertTrue( $this->in_publish_filter, 'Not in filter' );
@@ -161,7 +161,7 @@ class WP_Test_Publicize extends WP_UnitTestCase {
 		}
 	}
 
-	function set_post_flags_check( $flags, $post ) {
+	public function set_post_flags_check( $flags, $post ) {
 		if ( $flags['publicize_post'] ) {
 			$this->publicized_post_id = $post->ID;
 		}
@@ -169,7 +169,7 @@ class WP_Test_Publicize extends WP_UnitTestCase {
 		return $flags;
 	}
 
-	function prevent_publicize_post( $should_publicize, $post ) {
+	public function prevent_publicize_post() {
 		return false;
 	}