+++ projects/js-packages/connection/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ### This is a list detailing changes for the Jetpack RNA Connection Component releases.
 
+## 0.17.9 - 2022-04-19
+### Added
+- Add connected plugins list to the initial state and add action to update it when needed
+- Allow secondary users to connect from My Jetpack
+
+### Changed
+- Do not hard disconnect Jetpack on deactivation
+
+### Fixed
+- Avoid Warning in React PropTypes
+- do not display warning for secondary users if connection has an owner
+
+## 0.17.8 - 2022-04-12
+### Changed
+- Updated package dependencies.
+
 ## 0.17.7 - 2022-04-05
 ### Added
 - Add skipUserConnection option to connectScreen component