+++ projects/packages/connection/CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.38.0] - 2022-04-19
+### Added
+- Added list of connected list to the connection initial state
+- Add token lock functionality.
+
+### Changed
+- PHPCS: Fix `WordPress.Security.ValidatedSanitizedInput`
+
+### Deprecated
+- Deprecated in-place connection.
+
+## [1.37.6] - 2022-04-12
+### Changed
+- Updated package dependencies.
+
 ## [1.37.5] - 2022-04-06
 ### Removed
 - Removed tracking dependency.
@@ -545,6 +560,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Separate the connection library into its own package.
 
+[1.38.0]: https://github.com/Automattic/jetpack-connection/compare/v1.37.6...v1.38.0
+[1.37.6]: https://github.com/Automattic/jetpack-connection/compare/v1.37.5...v1.37.6
 [1.37.5]: https://github.com/Automattic/jetpack-connection/compare/v1.37.4...v1.37.5
 [1.37.4]: https://github.com/Automattic/jetpack-connection/compare/v1.37.3...v1.37.4
 [1.37.3]: https://github.com/Automattic/jetpack-connection/compare/v1.37.2...v1.37.3